backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
motor>=3.6.0
pymongo>=4.10.0
python-dotenv>=1.0.1
pydantic>=2.10.0
python-multipart>=0.0.9
websockets>=12.0
dnspython>=2.6.0
PyJWT>=2.8.0
cryptography>=42.0.0
httpx>=0.27.0
certifi>=2024.0.0
